+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - pipe to readme
+### Removed
+- create createDateWithMinutesFromNow
 
 ## [0.3.0] - 2018-11-14
 ### Added